requirements.txt
flask==1.1.1
Flask-RESTful==0.3.7
psutil
requests
pyjwt
